Refactor raw client & server
