+++ raw/common.go
@@ -1,0 +1,15 @@
+package raw
+
+import (
+	"encoding/binary"
+	"net"
+)
+
+func GetBodySize(conn *net.TCPConn) (uint64, error) {
+	var body_size uint64 = 0
+	err := binary.Read(conn, binary.BigEndian, body_size)
+	if err != nil {
+		return 0, err
+	}
+	return body_size, nil
+}
